requirements.txt
fastapi
uvicorn
python-multipart
jinja2
aiofiles
python-jose
requests
pycryptodome
SQLAlchemy
zmq
